+++ server/cmd/bbsd/main.go
@@ -18,8 +18,8 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/benedictemarty/bbsoric/server/internal/bbs"
 	"github.com/benedictemarty/bbsoric/internal/content"
+	"github.com/benedictemarty/bbsoric/server/internal/bbs"
 	"github.com/benedictemarty/bbsoric/server/internal/server"
 	"github.com/benedictemarty/bbsoric/server/internal/user"
 )